Revert "Allow visible_to check to be skipped for certain allow-listed packages (#6798)"
This reverts commit 6aa58421cc3d749c7344360844d472cc234cbb28.
