+++ core/packages/PackageDB.cc
@@ -245,10 +245,6 @@ const std::string_view PackageDB::errorHint() const {
     return errorHint_;
 }
 
-bool PackageDB::skipImportVisibilityCheckFor(core::NameRef mangledName) const {
-    return absl::c_find(skipImportVisibilityCheckFor_, mangledName) != skipImportVisibilityCheckFor_.end();
-}
-
 PackageDB PackageDB::deepCopy() const {
     ENFORCE(frozen);
     PackageDB result;
