+++ main/pipeline/pipeline.cc
@@ -680,10 +680,10 @@ vector<ast::ParsedFile> package(core::GlobalState &gs, vector<ast::ParsedFile> w
         {
             core::UnfreezeNameTable unfreezeToEnterPackagerOptionsGS(gs);
             core::packages::UnfreezePackages unfreezeToEnterPackagerOptionsPackageDB = gs.unfreezePackages();
-            gs.setPackagerOptions(
-                opts.secondaryTestPackageNamespaces, opts.extraPackageFilesDirectoryUnderscorePrefixes,
-                opts.extraPackageFilesDirectorySlashPrefixes, opts.packageSkipRBIExportEnforcementDirs,
-                opts.skipPackageImportVisibilityCheckFor, opts.stripePackagesHint);
+            gs.setPackagerOptions(opts.secondaryTestPackageNamespaces,
+                                  opts.extraPackageFilesDirectoryUnderscorePrefixes,
+                                  opts.extraPackageFilesDirectorySlashPrefixes,
+                                  opts.packageSkipRBIExportEnforcementDirs, opts.stripePackagesHint);
         }
         what = packager::Packager::run(gs, workers, move(what));
         if (opts.print.Packager.enabled) {
