--- test/cli/package-skip-import-visibility-check-for/__package.rb
@@ -1,6 +1,0 @@
-# typed: strict
-
-class Project::Root < PackageSpec
-  import Project::A
-end
-
