--- test/cli/package-skip-import-visibility-check-for/a/__package.rb
@@ -1,5 +1,0 @@
-# typed: strict
-
-class Project::A < PackageSpec
-  visible_to Project::B
-end
