--- test/cli/package-skip-import-visibility-check-for/b/__package.rb
@@ -1,5 +1,0 @@
-# typed: strict
-
-class Project::B < PackageSpec
-  import Project::A
-end
