--- test/cli/package-skip-import-visibility-check-for/test.out
@@ -1,1 +1,0 @@
-No errors! Great job.
