+++ test/testdata/packager/visibility/foo/__package.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 # typed: strict
 # enable-packager: true
-# skip-package-import-visibility-check-for: SkipCheck::For
 
 class Foo < PackageSpec
   visible_to Bar
