--- test/testdata/packager/visibility/skip_check/__package.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-# typed: strict
-# enable-packager: true
-
-class SkipCheck::For < PackageSpec
-  import Foo
-end
